third_party/blink/web_tests/external/wpt/css/css-backgrounds/parsing/background-position-x-computed-expected.txt

This is a testharness.js-based test.
PASS Property background-position-x value 'center' computes to '50%'
PASS Property background-position-x value 'left' computes to '0%'
PASS Property background-position-x value 'right' computes to '100%'
FAIL Property background-position-x value 'x-start' computes to 'x-start' assert_equals: expected "x-start" but got "0%"
FAIL Property background-position-x value 'x-end' computes to 'x-end' assert_equals: expected "x-end" but got "0%"
PASS Property background-position-x value '-20%' computes to '-20%'
PASS Property background-position-x value '10px' computes to '10px'
PASS Property background-position-x value '0.5em' computes to '20px'
PASS Property background-position-x value 'calc(10px - 0.5em)' computes to '-10px'
FAIL Property background-position-x value 'left -20%' computes to '-20%' assert_equals: expected "-20%" but got "0%"
FAIL Property background-position-x value 'right -10px' computes to 'calc(100% + 10px)' assert_equals: expected "calc(100% + 10px)" but got "0%"
PASS Property background-position-x value '-20%, 10px' computes to '-20%'
PASS Property background-position-x value 'center, left, right' computes to '50%'
FAIL Property background-position-x value '0.5em, x-start, x-end' computes to '20px' assert_equals: expected "20px" but got "0%"
PASS Property background-position-x value 'calc(10px - 0.5em), -20%, 10px' computes to '-10px'
FAIL Property background-position-x value 'calc(10px - 0.5em), left -20%, right 10px' computes to '-10px' assert_equals: expected "-10px" but got "0%"
Harness: the test ran to completion.
